+++ src/validations/css/prettycss.js
@@ -1,7 +1,7 @@
 import trim from 'lodash/trim';
 import endsWith from 'lodash/endsWith';
 import Validator from '../Validator';
-import retryingFailedImports from '../retryingFailedImports';
+import retryingFailedImports from '../../util/retryingFailedImports';
 
 const RADIAL_GRADIENT_EXPR =
   /^(?:(?:-(?:ms|moz|o|webkit)-)?radial-gradient|-webkit-gradient)/;
