+++ src/validations/html/htmllint.js
@@ -2,7 +2,7 @@ import clone from 'lodash/clone';
 import defaults from 'lodash/defaults';
 import reduce from 'lodash/reduce';
 import Validator from '../Validator';
-import retryingFailedImports from '../retryingFailedImports';
+import retryingFailedImports from '../../util/retryingFailedImports';
 
 const errorMap = {
   E001: (error) => {
